+++ frontend/src/app/trees/forests/tree-guidelines/tree-cutting-dates.component.spec.ts
@@ -5,6 +5,7 @@ import { TreeCuttingDatesComponent } from './tree-cutting-dates.component';
 import { Pipe, PipeTransform } from '@angular/core';
 import { LineBreakFormatterPipe } from '../../../_pipes/line-break-formatter.pipe';
 import * as sinon from 'sinon';
+import { TreeDistrictsUtilService } from './tree-districts-util.service';
 
 describe('TreeCuttingDatesComponent', () => {
   let component: TreeCuttingDatesComponent;
@@ -13,7 +14,8 @@ describe('TreeCuttingDatesComponent', () => {
   beforeEach(
     async(() => {
       TestBed.configureTestingModule({
-        declarations: [TreeCuttingDatesComponent, FilterPipe, LineBreakFormatterPipe]
+        declarations: [TreeCuttingDatesComponent, FilterPipe, LineBreakFormatterPipe],
+        providers: [TreeDistrictsUtilService]
       }).compileComponents();
     })
   );
