+++ frontend/src/app/trees/forests/tree-guidelines/tree-locations/tree-locations-allowed/tree-locations-allowed.component.spec.ts
@@ -3,9 +3,8 @@ import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { TreeLocationsAllowedComponent } from './tree-locations-allowed.component';
 import { forest } from '../../../../_mocks/forest.mock';
-import { TreesService } from '../../../../_services/trees.service';
-import { MockBackend } from '@angular/http/testing';
 import { UtilService } from '../../../../../_services/util.service';
+import { TreeDistrictsUtilService } from '../../tree-districts-util.service';
 
 describe('TreeLocationsAllowedComponent', () => {
   let component: TreeLocationsAllowedComponent;
@@ -16,7 +15,7 @@ describe('TreeLocationsAllowedComponent', () => {
       TestBed.configureTestingModule({
         declarations: [TreeLocationsAllowedComponent],
         schemas: [NO_ERRORS_SCHEMA],
-        providers: [{ provide: UtilService }]
+        providers: [TreeDistrictsUtilService, { provide: UtilService }]
       }).compileComponents();
     })
   );
@@ -35,12 +34,12 @@ describe('TreeLocationsAllowedComponent', () => {
   it('should get allowed districts', () => {
     component.forest = forest;
     component.populateDistricts();
-    expect(component.allowedDistricts[0].childNodes[0].id).toEqual(17);
+    expect(component.allowedDistricts[0].locations[0].id).toEqual(17);
   });
 
   it('should get run allowed districts on changes', () => {
     component.forest = forest;
     component.ngOnChanges();
-    expect(component.allowedDistricts[0].childNodes[0].id).toEqual(17);
+    expect(component.allowedDistricts[0].locations[0].id).toEqual(17);
   });
 });
