+++ frontend/src/app/trees/forests/tree-guidelines/tree-locations/tree-locations-allowed/tree-locations-allowed.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input, OnChanges, SimpleChanges } from '@angular/core';
 import { UtilService } from '../../../../../_services/util.service';
+import { TreeDistrictsUtilService } from '../../tree-districts-util.service';
 
 @Component({
   selector: 'app-tree-locations-allowed',
@@ -8,29 +9,16 @@ import { UtilService } from '../../../../../_services/util.service';
 export class TreeLocationsAllowedComponent implements OnChanges {
   @Input() forest: any;
   allowedDistricts: any = [];
-  constructor(public util: UtilService) {}
+  constructor(public util: UtilService, public districtUtil: TreeDistrictsUtilService) {}
 
   populateDistricts() {
     const locations = this.forest.locations;
-    const districts = Object.create(null);
-
-    locations.forEach(location => (districts[location.district] = { ...location, childNodes: [] }));
-
-    delete districts.null;
-
-    const districtsTree = [];
-
-    locations.forEach(location => {
-      if (location.district) {
-        districts[location.district].childNodes.push(location);
-      } else {
-        districtsTree.push(districts[location]);
-      }
-    });
+    const districts = this.districtUtil.mapLocationsToDistricts(locations);
+    this.districtUtil.reduceLocations(locations, districts);
 
     for (const key of Object.keys(districts)) {
       const district = districts[key];
-      if (district.childNodes.length && district.allowed) {
+      if (district.locations.length && district.allowed) {
         this.allowedDistricts.push(district);
       }
     }
