+++ frontend/src/app/trees/forests/tree-guidelines/tree-locations/tree-locations-prohibited/tree-locations-prohibited.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input, OnChanges, SimpleChanges } from '@angular/core';
 import { FilterPipe } from '../../../../../_pipes/filter.pipe';
+import { TreeDistrictsUtilService } from '../../tree-districts-util.service';
 
 @Component({
   selector: 'app-tree-locations-prohibited',
@@ -10,29 +11,16 @@ export class TreeLocationsProhibitedComponent implements OnChanges {
   @Input() forest: any;
   prohibitedDistricts: any = [];
 
-  constructor(private filter: FilterPipe) {}
+  constructor(private filter: FilterPipe, public districtUtil: TreeDistrictsUtilService) {}
 
   populateDistricts() {
     const locations = this.forest.locations;
-    const districts = Object.create(null);
-
-    locations.forEach(location => (districts[location.district] = { ...location, childNodes: [] }));
-
-    delete districts.null;
-
-    const districtsTree = [];
-
-    locations.forEach(location => {
-      if (location.district) {
-        districts[location.district].childNodes.push(location);
-      } else {
-        districtsTree.push(districts[location]);
-      }
-    });
+    const districts = this.districtUtil.mapLocationsToDistricts(locations);
+    this.districtUtil.reduceLocations(locations, districts);
 
     for (const key of Object.keys(districts)) {
       const district = districts[key];
-      if (district.childNodes.length && !district.allowed) {
+      if (district.locations.length && !district.allowed) {
         this.prohibitedDistricts.push(district);
       }
     }
